[11.x] Fix the chunk method to an integer type in the splitIn method

+++ src/Illuminate/Collections/LazyCollection.php
@@ -1317,7 +1317,7 @@ class LazyCollection implements CanBeEscapedWhenCastToString, Enumerable
      */
     public function splitIn($numberOfGroups)
     {
-        return $this->chunk(ceil($this->count() / $numberOfGroups));
+        return $this->chunk((int) ceil($this->count() / $numberOfGroups));
     }
 
     /**
